fix: unblock check_authority_reconcile.py (file-level LLM marker misfire)
tools/check_authority_reconcile.py's declared-vs-detected authority check
failed: "package 'reconciliation': declared CODE-ONLY but 2/145 roadmap-AC
test(s) are LLM tests (detected CODE-LED)". Root cause:
common/meta/extension/authority_classifier.py classifies an entire test
FILE as an LLM test if it contains any of a handful of literal markers
(including the substring "stream_ai_json") ANYWHERE in the file — a
file-level heuristic, not per-function. The new AC-reconciliation.1803.1/.2
tests lived in test_ai_reconciliation.py alongside pre-existing
ai_semantic_score tests that legitimately mock stream_ai_json directly, so
the whole file (and therefore both new roadmap-registered tests, even
though their own bodies just assert deterministic gating/arithmetic) got
classified as LLM.

Moved the 2 hybrid/feature-flag tests into a new file
(test_reconciliation_hybrid_scoring.py) and changed their mock target from
the low-level stream_ai_json/accumulate_stream to ai_semantic_score itself
(the function calculate_match_score actually calls) — removes the marker
string from both the new file and the tests' own scope, and is arguably a
more correctly-scoped mock (these ACs are about the 60-84 band gate and the
0.7/0.3 blend, not about how ai_semantic_score gets its answer).

Verified: check_authority_reconcile.py now reports reconciliation
code=145 llm=0 share=0.0% (was 143 llm=2); full pytest tests/tooling/ —
2138 passed; ruff check/format clean.
